+++ lib/domain/models/media_collection_model.dart
@@ -37,10 +37,10 @@ class MediaCollection {
 
   /// Remove duplicates from the collection and return the count removed
   int removeDuplicates() => grouping.removeDuplicates(_media);
-
-  /// Async version of removeDuplicates for better performance
-  /// Uses streaming hash calculation to avoid loading entire files into memory
-  Future<int> removeDuplicatesAsync() => grouping.removeDuplicatesAsync(_media);
+  /// Ultra-fast async version of removeDuplicates for massive performance gains
+  /// Uses parallel processing to dramatically improve performance on large collections
+  Future<int> removeDuplicatesAsyncOptimized() => 
+      grouping.removeDuplicatesAsyncOptimized(_media);
 
   /// Remove "extra" files (edited versions) and return the count removed
   int removeExtras() => extras.removeExtras(_media);